+++ retroshare-gui/src/gui/connect/PGPKeyDialog.cpp
@@ -318,7 +318,7 @@ void PGPKeyDialog::applyDialog()
     }
 
     //check the GPG trustlvl
-    if(ui.trustlevel_CB->currentIndex() != detail.trustLvl)
+    if(ui.trustlevel_CB->currentIndex() != (int)detail.trustLvl)
         rsPeers->trustGPGCertificate(pgpId, ui.trustlevel_CB->currentIndex());
 
     uint32_t max_upload_speed = ui.maxUploadSpeed_SB->value() ;
